Set duration as double instead of float.

+++ editor/actions/fade.h
@@ -59,7 +59,7 @@ private:
     void init();
 
     Fade::Type mFadeType;
-    float mDuration;
+    double mDuration;
 
 };
